remove ruby2/syck hack, fixes #1
rather than hack in a syck dependency for ruby 2, just have different gems
for different rubies. while it's possible to make it work the old way
(need to activate the syck gemspec at runtime), it circumvents bundler and
isn't very polite (not in Gemfile.lock, etc.).

note that we don't just add syck for 1.9, because 1. it's already included
in ruby and 2. the syck gem doesn't install on 1.9

test plan:
1. tests should pass on 1.9, 2.0, and 2.1
2. you should be able to use i18nema(19) in ruby 1.9, 2.0, and 2.1 apps
   without having to explictly put syck in your Gemfile
   e.g. in a rails app, add it to your Gemfile and confirm that this works
   in a rails console:
    > I18nema::Backend.new.direct_lookup
   => {}

Change-Id: I2abe507c03c4738a7e9e0c52eae32a50e8556c0b